moved swap off check in platform module
