+++ pkg/util/constants.go
@@ -9,6 +9,7 @@ import (
 var RequiredPorts []string
 var PortErr string
 var ProcessesList []string //Kubernetes clusters processes list
+var SwapOffDisabled bool   //If this is true the swapOff functionality will be disabled.
 
 const (
